fix : UPDATE move from step3 package to production package
