+++ src/test/java/racingcar/controller/RacingTest.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
