+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,8 +7,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
-
-import racingcar.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
