+++ src/test/java/racingcar/view/ResultTest.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,7 +8,6 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.util.stream.Stream;
 
 import racingcar.domain.Car;
-import racingcar.view.ResultView;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -29,5 +28,4 @@ public class ResultTest {
                 Arguments.of(false, "")
         );
     }
-
 }
